Move lower_compute_at_map to compute_at_map.
